memory mgmt
Former-commit-id: a40fbe420f9909b07931817c1069192618287b44

+++ os/format.h
@@ -1,8 +1,10 @@
 
 #define FORMAT_MAX_SIZE 1024
 
+#include <Yap.h>
+
 typedef struct {
-  Int filler;
+  intptr_t filler;
   /* character to dump */
   int phys;
   /* position in buffer */
